==> 02-data.Rmd <==
# Data sources

[DOHMH New York City Restaurant Inspection Results](https://data.cityofnewyork.us/Health/DOHMH-New-York-City-Restaurant-Inspection-Results/43nn-pn8j)

## Who is responsible for collecting the data?

NYC government's health department conducts unexpected inspections for restaurants in New York City at least once a year. During COVID-19 before August 14 2021, the inspection was stopped but now the NYC Health Department is conducting full inspections, including issuing letter grades and summonses subject to fines.
 
## Why and How is the data collected?

All restaurants are required to be in compliance with NYS and NYC Food Safety Regulations, found in New York City Health Code Article 81. By inspection, health department want to check that restaurants comply with food safety rules. Violations of food safety rules have point values for each item, and total score of restaurant results in a specific letter grade. The lower the score, the better the grade.

## Basic information about the dataset

The dataset contains details on violation citations against restaurants from every inspection conducted by the NYC government.It includes inspection results for active restaurants for the last three years. Inactive restaurants and any violations cited during the inspection that were dismissed during adjudication are excluded. Here is the basic information of dataset and each variable.

* Update Frequency : Daily
* Num of records : 383,199
* Agency: Department of Health and Mental Hygiene (DOHMH)
* Each row is a : Restaurant Citations

```{r, echo=FALSE, message=FALSE, warning=FALSE}
library(tidyverse)
library(reactable)
info <- read.csv("data/variable_info.csv")
reactable(info)
```

## Issues

Records are also included for each restaurant that has applied for a permit but has not yet been inspected and for inspections resulting in no violations. Establishments with inspection date of 1/1/1900 are new establishments that have not yet received an inspection. This records can make errors when analyzing the data. So we will handle this records in 'Data transformation' Chapter.
