==> 03-cleaning.Rmd <==
# Data transformation

```{r message = FALSE}
library(ggplot2)
library(tidyverse)
library(dplyr)
library(tidyr)

```

Let us begin by loading the dataset

```{r, echo=FALSE}
df <- readRDS(file='data/DOHMH_New_York_City_Restaurant_Inspection_Results.Rda')

head(df)
```

We will first drop the columns for which the creator has not provided any information as well as those columns which we believe are not useful for drawing insights.

```{r, echo=FALSE}
df <- subset(df, select = -c(BUILDING, PHONE, RECORD.DATE, GRADE.DATE, Community.Board, Council.District, Census.Tract, BIN, BBL, NTA) )

head(df)
```

Let us rename each column to its shorter version for easier access

```{r, echo=FALSE}
names(df) <- c('camis', 'dba', 'boro', 'street', 'zip', 'cuisine', 'insp.date', 'action', 'vio.code', 'vio.desc', 'flag', 'score', 'grade', 'insp.type', 'latitude', 'longitude')
```

We will separate out inspection type into two columns denoting inspection program and inspection type

```{r, echo=FALSE}
df <- df %>%
  separate(insp.type, c("insp.prog", "insp.type"), sep = " / ")

head(df)
```

Consider the unique values in borough
```{r, echo=FALSE}
table(df$boro)
```

'0' borough is vague, we'll drop these rows: 
```{r, echo=FALSE}
print(paste0('There are ', count(df[df$boro == '0',]), ' rows with 0 as the borough'))
```

```{r, echo=FALSE}
df <- df[df$boro!='0',]
```

Consider the unique values in grade column:

```{r, echo=FALSE}
table(df$grade)
```


As per the data dictionary, grade value 'G' is not defined. Since only one row contains this, we can get rid of this row.

```{r}
df <- df[!(df$grade %in% c('G')),]
```

```{r, echo=FALSE}
sapply(df, class)
```


```{r, echo=FALSE}
factor_cols = c('camis', 'dba', 'boro', 'zip', 'cuisine', 'action', 'vio.code', 'vio.desc', 'flag', 'grade', 'insp.prog', 'insp.type')

for (col in factor_cols) {
  df[[col]] <- as.factor(df[[col]])
}

date_cols = c('insp.date')

for (col in date_cols) {
  df[[col]] <- as.Date(df[[col]] , format="%m/%d/%Y") 
}

sapply(df,class)
```


We will now drop the duplicate rows

```{r, echo=FALSE}
print(paste0("Number of rows originally: ", count(df)))

df <- distinct(df)

print(paste0("Number of rows after removing duplicates: ", count(df)))
```

We will replace lengthy categories in 'action' by keywords

```{r, include = FALSE}
df <- subset(df, !(action %in% c('05D','10F')))
```

```{r, include = FALSE}
df$action <- factor(df$action)
```

```{r}
table(df$action)
```

```{r, echo=FALSE}
levels(df$action) <- c('closed', 'reclosed', 'reopened', 'no violations', 'violations')
```

```{r}
table(df$action)
```

Let us now save this cleaned dataset
```{r, echo=FALSE}
df <- df %>%
  janitor::clean_names()

saveRDS(df, file='data/clean/post_cleaning.Rda')
```
