==> 04-missing.Rmd <==
# Missing values


```{r, echo=FALSE}
library(ggplot2)
library(tidyverse)
library(dplyr)
library(tidyr)
library(patchwork)
library(lubridate)
```


```{r, echo=FALSE}
# create a function (include it in the Rmd file)
# function have a functional parameter 'percent' to show either counts or percent (default : FALSE)
plot_missing <- function(x, percent = FALSE){
  
  if(percent == TRUE){
    missing_patterns <- data.frame(is.na(x)) %>%
      group_by_all() %>%
      count(name = "count", sort = TRUE) %>%
      ungroup() %>% 
      rownames_to_column("id")
    
    bypattern <- missing_patterns[,c('id','count')] %>% 
      mutate(count_per = count/sum(count)) %>% 
      select("id","count_per")
    bycolumn <- data.frame(colMeans(is.na(x))) %>% 
      rownames_to_column("key")
    colnames(bycolumn)[2] <- 'count_per'
    
    
    missing_patterns <- missing_patterns[,!names(missing_patterns) %in% c('count')] 
    
    missing_patterns <- missing_patterns %>% 
      gather(key, value, -id) %>% 
      mutate(pattern = ifelse(value != 0, "yes", "no"))
    
    missing_patterns <- merge(missing_patterns, bycolumn, by = "key")
    
    missing_patterns <- missing_patterns %>% 
      mutate(pattern2 = ifelse(value != 0, 1, 0))
    
    complete <- missing_patterns %>% 
      group_by(id) %>% 
      summarize(count = sum(pattern2)) %>% 
      mutate(complete = ifelse(count == 0, 1, 0)) %>% 
      select("id", "complete")
    
    missing_patterns <- merge(missing_patterns, complete, by = "id")
    bypattern <- merge(bypattern, complete, by = "id")
    bypattern$complete <- ifelse(bypattern$complete == 1, "yes", "no")
    
    
    missing_patterns <- missing_patterns %>% 
      mutate(patterncomplete = ifelse(complete == 1, "complete", pattern))
    
    bypattern$count_per <- bypattern$count_per*100
    bycolumn$count_per <- bycolumn$count_per*100
    
    missing_patterns <- missing_patterns[order(as.numeric(as.character(missing_patterns$id))),]
    bypattern <- bypattern[order(as.numeric(as.character(bypattern$id))),]
    
    p1 <- ggplot(missing_patterns, aes(x = fct_reorder(key, -count_per),  y=reorder(id, sort(as.numeric(id), decreasing = TRUE)), fill = pattern, alpha = patterncomplete)) +
      geom_tile(color = "white") +
      scale_fill_viridis_d() + 
      theme_bw() + guides(fill = "none", alpha = "none") +
      xlab("variable") +
      ylab("missing pattern") +
      scale_alpha_manual(values = c(0.5, 1, 1))
    
    p2 <- ggplot(bycolumn, aes(x=fct_reorder(key, -count_per), y=count_per)) + 
      geom_bar(stat = "identity") +
      theme(axis.title.x = element_blank()) +
      ylab("% rows missing:") +
      coord_cartesian(ylim = c(0, 100)) 
    
    p3 <- ggplot(bypattern, aes(x=count_per,  y=reorder(id, sort(as.numeric(id), decreasing = TRUE)), alpha = complete)) + 
      geom_bar(stat = "identity") +
      theme(axis.title.y = element_blank()) +
      xlab("% rows") + guides(alpha = "none") +
      scale_alpha_manual(values = c(1, 0.5)) +
      coord_cartesian(xlim = c(0, 100)) 
    
    wrap_plots(p2, plot_spacer(), p1, p3, nrow = 2, widths = c(1.5, 0.5), heights = c(0.5, 1.5)) 
    
    
  }else{
    missing_patterns <- data.frame(is.na(x)) %>%
      group_by_all() %>%
      count(name = "count", sort = TRUE) %>%
      ungroup() %>% 
      rownames_to_column("id")
    
    bypattern <- missing_patterns[,c('id','count')] 
    bycolumn <- data.frame(colSums(is.na(x))) %>% 
      rownames_to_column("key")
    colnames(bycolumn)[2] <- 'count'
    
    missing_patterns <- missing_patterns[,!names(missing_patterns) %in% c('count')] 
    
    missing_patterns <- missing_patterns %>% 
      gather(key, value, -id) %>% 
      mutate(pattern = ifelse(value != 0, "yes", "no"))
    
    missing_patterns <- merge(missing_patterns, bycolumn, by = "key")
    
    missing_patterns <- missing_patterns %>% 
      mutate(pattern2 = ifelse(value != 0, 1, 0))
    
    complete <- missing_patterns %>% 
      group_by(id) %>% 
      summarize(count = sum(pattern2)) %>% 
      mutate(complete = ifelse(count == 0, 1, 0)) %>% 
      select("id", "complete")
    
    missing_patterns <- merge(missing_patterns, complete, by = "id")
    bypattern <- merge(bypattern, complete, by = "id")
    bypattern$complete <- ifelse(bypattern$complete == 1, "yes", "no")
    
    
    missing_patterns <- missing_patterns %>% 
      mutate(patterncomplete = ifelse(complete == 1, "complete", pattern))
    
    
    missing_patterns <- missing_patterns[order(as.numeric(as.character(missing_patterns$id))),]
    bypattern <- bypattern[order(as.numeric(as.character(bypattern$id))),]
    
    p1 <- ggplot(missing_patterns, aes(x = fct_reorder(key, -count), y=reorder(id, sort(as.numeric(id), decreasing = TRUE)), fill = pattern, alpha = patterncomplete)) +
      geom_tile(color = "white") +
      scale_fill_viridis_d() + 
      theme_bw() + guides(fill = "none", alpha = "none") +
      xlab("variable") +
      ylab("missing pattern") +
      scale_alpha_manual(values = c(0.5, 1, 1))
    
    p2 <- ggplot(bycolumn, aes(x=fct_reorder(key, -count), y=count)) + 
      geom_bar(stat = "identity") +
      theme(axis.title.x = element_blank()) +
      ylab("num rows missing:")
    
    p3 <- ggplot(bypattern, aes(x=count, y=reorder(id, sort(as.numeric(id), decreasing = TRUE)), alpha = complete)) + 
      geom_bar(stat = "identity") +
      theme(axis.title.y = element_blank()) +
      xlab("row count") + guides(alpha = "none") +
      scale_alpha_manual(values = c(1, 0.5))
    
    wrap_plots(p2, plot_spacer(), p1, p3, nrow = 2, widths = c(1.5, 0.5), heights = c(0.5, 1.5)) 
    
    
    
    
  }
}

```

```{r, echo=FALSE}
# load the cleaned dataset
df <- readRDS(file='data/clean/post_cleaning2.Rda') %>%
  select(-street, -longitude, -latitude)

# shorten the column names so they don't overlap in the plot
names(df) <- abbreviate(names(df), minlength=5)
```


Let us analyze the missing patterns in our data frame

```{r, echo=FALSE, fig.height=7, fig.width=12}
# find missing patterns in dataset (counts)
plot_missing(df, percent = FALSE)

# find missing patterns in dataset (percents)
plot_missing(df, percent = TRUE)
```


According to the output of plot_missing function, we can confirm that there are a total of 29 missing patterns in our dataset. But if we look at plots in detail, the most common pattern is the one that has no missing values, followed by the pattern with missing values only in the 'grade' column. These two cases account for almost all of the most frequent missing patterns. We can also infer from the upper side plot that 'grade' has about 50% of its missing values and all other columns have relatively much fewer proportions of missing values.

Since there are so many rows with missing values in 'grade', dropping these rows would not make sense. Instead, we will try to impute the 'grade' column values by finding exciting patterns in the dataset in further exploration. 


Let us now try to see how many rows have a given number of missing values.
```{r, echo=FALSE, fig.height=7, fig.width=12}
t <- rowSums(is.na(df)) %>%
  sort(decreasing = TRUE) 
  
ggplot() + 
  aes(t) + 
  geom_histogram(binwidth=1, colour="black", fill="lightblue") +
  xlab('Number of columns with missing values') +
  ylab('Number of rows') +
  ggtitle('Number of rows with given number of missing values') +
  scale_x_continuous(breaks = round(seq(min(t), max(t), by = 1),1)) 
```

We can see that most of the rows in our dataset either have no missing values or have exactly one missing value. The reason for this is evident from the missing-patterns plot we showed above; since the column 'grade' has so many missing values. This will be fixed when we impute the values for 'grade'. We will consider dropping rows with more than five missing values; this is reasonable as these rows are relatively less in quantity.


To better understand the pattern of missing values, we have to check for any placeholders for missing values. A nice strategy for this is to look for outliers.

Let us try to see when most of the inspections were conducted.

```{r, echo=FALSE, fig.height=7, fig.width=12}
t <- as.numeric(format(df$insp_d,"%Y"))

ggplot() + aes(t) + geom_boxplot() + xlab('Year of inspection')
```


Here, we can see that the inspection date with year = 1900 is an outlier and possibly a placeholder for missing values. We confirm below by analyzing the missing patterns of the subset of our dataset with the inspection year 1900.

```{r, echo=FALSE, fig.height=7, fig.width=12}
t <- subset(df, df$insp_d>='1900-01-01' & df$insp_d<='1900-12-31')[, !colnames(df) %in% c('insp_d')]


plot_missing(t, percent = TRUE)
```

We can see that most of the data in rows with the inspection year equal to 1900 is missing. Therefore, we will consider dropping these rows.

Let us try to understand if there's any relation between missing-patterns and borough.

```{r}
unique(df$boro)
```

```{r, echo=FALSE}
t <- subset(df, df$boro=='Manhattan')[, !colnames(df) %in% c('boro')]
plot_missing(t, percent = TRUE) + ggtitle('Manhattan borough')
```


```{r, echo=FALSE, fig.height=7, fig.width=12}
t <- subset(df, df$boro=='Queens')[, !colnames(df) %in% c('boro')]
plot_missing(t, percent = TRUE) + ggtitle('Queens borough')
```

```{r, echo=FALSE, fig.height=7, fig.width=12}
t <- subset(df, df$boro=='Brooklyn')[, !colnames(df) %in% c('boro')]
plot_missing(t, percent = TRUE) + ggtitle('Brooklyn borough')
```

```{r, echo=FALSE, fig.height=7, fig.width=12}
t <- subset(df, df$boro=='Staten Island')
plot_missing(t, percent = TRUE) + ggtitle('Island borough')
```

```{r, echo=FALSE, fig.height=7, fig.width=12}
t <- subset(df, df$boro=='Bronx')
plot_missing(t, percent = TRUE) + ggtitle('Bronx borough')
```

We can see that none of these seem to display any relation between missing-patterns and boroughs, which is worth investigating further.


## Transformation

We drop the rows that have 1900 as their inspection year

```{r, echo=FALSE}
df <- readRDS(file='data/clean/post_cleaning2.Rda')

df <- df[!(df$insp_date>='1900-01-01' & df$insp_date<='1900-12-31'),]

```


```{r}
head(df)
```


Further, consider 

```{r}
t <- df

t <- t %>%
  count(insp_year = year(insp_date)) %>%
  slice(1:10) 

t$insp_year <- as.factor(as.character(t$insp_year))

t
```
We can see that there are very few observations from the years before 2016. In order to draw meaningful insights, we shall drop all rows with inspection year less than 2016 as we have very less data from this time period. Hence, we'll analyse only the subset with inspection year from 2016 to 2021.

```{r, echo=FALSE}
df <- df[df$insp_date>='2016-01-01',]

```


Let us now save this dataset for further analysis

```{r, echo=FALSE}
head(df)

saveRDS(df, file='data/clean/post_missing.Rda')
```
